Update README.md
Removed the suggestion of using EditThisCookie, as it maliciously collects large amounts of user data that are completely unrelated to its function, like your **location**, as well as your browsing history and other such metrics. I instead replaced it with the AwesomeCookieManager, which is slightly less user-friendly but doesn't steal your data, so I'd say it's a pretty decent trade-off.